pyproject.toml
[project]
name = "cmrn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "ipykernel>=6.29.5",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "transformers[torch]>=4.51.3",
]
